ingest/requirements.txt
requests>=2.31.0
google-cloud-bigquery>=3.11.0
google-auth>=2.23.0
